Used component composition in AsideMenu.jsx to not pass so many props to this component from MainView.jsx
